Update gap.c
Starting off with a Mac Address, then go into your settings and turn Bluetooth off and back on, and you'll have a new Mac Address.  As of right now, can only connect to second address.  Will update when I'm ready to write new code to allow for more changes, and the ability to connect to all addresses.